Create Homework1.R
STAT 202A HOMEWORK 1:

Please implement the sweep operator and the Gauss-Jordan elimination (both the plain version and the vectorized version) in both R and Python. Please confirm that the sweep operator is a compact version of the Gauss-Jordan elimination using some concrete examples.

Optional 1: Please fortify your code by checking whether the input matrix is a squared matrix, and whether the input matrix is invertible (i.e., the pivot elements should not be too close to zero).

Optional 2: please write R and Python code for computing the determinant of a matrix using the sweep operator.

Optional 3: please write R and Python code for the reverse sweep operator, and confirm that it indeed reverses the sweep operator.